backend/pyproject.toml
[tool.poetry]
name = "djangorestapp"
version = "0.1.0"
description = ""
authors = ["liammcbride <mailmcbride56@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
